rewrite: Change association method
Before this patch, `thin_delegate` used a global variable in proc macro
to transfer definitions of traits and structs/enums to where it
generates impl of a trait.

This method is problematic. For example, it's not ensured that the same
process is used to invoke proc macros, even if it's in the same
compilation unit (crate). E.g. incremental builds like rust-analyzer.
See e.g. [1], [2], and [3] for more details.

We resolve this problem by using declarative macros to transfer the
data.

This CL updates fundamental structures of this crate and updates tests
as much as possible. Remaining tests will be updated in the next
patches.

[1] https://github.com/rust-lang/rust/issues/44034
[2] https://users.rust-lang.org/t/simple-state-in-procedural-macro/68204
[3] https://stackoverflow.com/questions/52910783/is-it-possible-to-store-state-within-rusts-procedural-macros