feat!: Switched default `delegate_fn_with_default_impl` from `false` to `true`
I was thinking about which is more natual, but I found that both is not
natural because there are conficting intuitions and user expectation:

A. Trait implementation doesn't require to implement trait function with
   default implementation.
B. `thin_delegate::fill_delegate` fills trait function by delegation.

So, it looks up to our dicision.

There are three reasons that we prefer `true` for the default value:

1. The description of default behavior is simpler: It fills trait
functions.

2. Default implementations can call trait functions that is necessary to
implement multiple times. For example, `std::io::Write::write_all()`
calls `write()` in a while loop. If a delegator is an enum,
using default implementation passes `match` multiple times, but
delegating outer method passes `match` once. It doesn't mean the latter
is more efficient, but it would be simpler.

3. There is a pattern that a trait provides default implementations for
all trait functions, e.g. `syn::visitor::Visitor`. In this case, trait
implementation is expected to implement trait functions that are
different from default, and it's not derived from other trait functions.

The downside of `true` would be only that it doesn't meet the
expectation A. So, we change the default value to `true`.